created seprate files for styles
