+++ src/components/enquiryForm/mobile/enquiryForm.jsx
@@ -1,56 +1,12 @@
-import { Box, Dialog,  IconButton, Slide, styled, useMediaQuery } from "@mui/material";
+import { Box, Dialog, IconButton, Slide, styled, useMediaQuery } from "@mui/material";
 import React, { useState } from "react";
 import { FormInput } from "../../widgets/formInput";
 import { FormButton } from "../../widgets/formButton";
 import CloseIcon from "@mui/icons-material/Close";
-
+import { Container, Header } from "./enquiryFormMobile.styles";
 const Transition = React.forwardRef(function Transition(props, ref) {
   return <Slide direction="up" ref={ref} {...props} />;
 });
-
-const Container = styled(Box)(({ theme }) => ({
-  background: `linear-gradient(${theme.palette.primary.main}, ${theme.palette.primary.main}80)`,
-  height: "100%",
-  padding: "1rem",
-  display: "flex",
-  flexDirection: "column",
-  gap: "0.2rem",
-
-  div: {
-    display: "flex",
-    gap: "1rem",
-  },
-
-  form: {
-    display: "flex",
-    flexDirection: "column",
-    gap: "0.875rem",
-  },
-}));
-
-const Header = styled(Box)(({ theme }) => ({
-  display: "flex",
-  justifyContent: "space-between",
-  div: {
-    display: "flex",
-    flexDirection: "column",
-    alignItem: "center",
-
-    h1: {
-      fontFamily: theme.typography.primaryTitle.fontFamily,
-      fontSize: "1.2rem",
-      margin: "unset",
-    },
-
-    h3: {
-      fontFamily: theme.typography.primaryTitle.fontFamily,
-      fontSize: "1rem",
-      fontWeight: 500,
-      margin: "unset",
-      marginBottom: "0.5rem",
-    },
-  },
-}));
 
 export const EnquiryForm = ({ name, open, setOpen }) => {
   const isVerySmallScreen = useMediaQuery("(max-width:360px)");
@@ -81,10 +37,25 @@ export const EnquiryForm = ({ name, open, setOpen }) => {
         </Header>
         <form>
           <div>
-            <FormInput label="Name" placeholder={"John doe"} value={form.name} onChange={(e) => setForm({ ...form, name: e.target.value })} />
-            <FormInput label="Phone" placeholder={"+91 9876543221"} value={form.phone} onChange={(e) => setForm({ ...form, phone: e.target.value })} />
+            <FormInput
+              label="Name"
+              placeholder={"John doe"}
+              value={form.name}
+              onChange={(e) => setForm({ ...form, name: e.target.value })}
+            />
+            <FormInput
+              label="Phone"
+              placeholder={"+91 9876543221"}
+              value={form.phone}
+              onChange={(e) => setForm({ ...form, phone: e.target.value })}
+            />
           </div>
-          <FormInput label="Email" placeholder={"example@exmaple.com"} value={form.email} onChange={(e) => setForm({ ...form, email: e.target.value })} />
+          <FormInput
+            label="Email"
+            placeholder={"example@exmaple.com"}
+            value={form.email}
+            onChange={(e) => setForm({ ...form, email: e.target.value })}
+          />
           <FormInput
             label="Leave a message"
             placeholder={"Please Leave your message here..."}
