+++ src/components/faqSection/faqSections.jsx
@@ -7,52 +7,17 @@ import { Box, Collapse, Typography, styled } from "@mui/material";
 import { PrimaryText, PrimarySubText, AnchorText } from "../common/typographies/typographies";
 import { useState } from "react";
 
-const Container = styled(Box)(({ theme }) => ({
-  background: theme.palette.faqSection.background,
-  paddingBottom: "2rem",
-  fontFamily: "'Aquawax', Courier, monospace;",
-}));
-
-const TitleText = styled(PrimaryText)(({ theme }) => ({
-  fontSize: "2.5rem",
-  paddingTop: "1.344rem",
-  paddingBottom: "0.625rem",
-  fontFamily: "'Aquawax', Courier, monospace;",
-  fontWeight: 500,
-  [theme.breakpoints.down("sm")]: {
-    fontSize: "1.563rem",
-  },
-}));
-
-const SubtitleText = styled(PrimarySubText)(({ theme }) => ({
-  fontFamily: "'Aquawax', Courier, monospace;",
-  fontSize: "0.875rem",
-  [theme.breakpoints.down("sm")]: {
-    paddingLeft: "2rem",
-    paddingRight: "2rem",
-  },
-}));
-
-const FaqContainer = styled(Box)(({ theme }) => ({
-  paddingLeft: "10%",
-  paddingRight: "10%",
-  width: "calc(100% - 20%)",
-  [theme.breakpoints.down("sm")]: {
-    paddingLeft: "5%",
-    paddingRight: "5%",
-    width: "calc(100% - 10%)",
-  },
-}));
-
-const ReadMore = styled(Typography)(({ theme }) => ({
-  color: theme.palette.buttonPrimary,
-  marginTop: "1rem",
-  display: "flex",
-  gap: "0.2rem",
-  justifyContent: "center",
-  alignItems: "center",
-  verticalAlign: "baseline",
-}));
+import {
+  Container,
+  TitleText,
+  SubtitleText,
+  FaqContainer,
+  ReadMore,
+  QuestionContainer,
+  QuestionTitleContainer,
+  QuestionTitle,
+  QuestionAnswer,
+} from "./faqSection.styles";
 
 export const FaqSection = () => {
   const [showMore, setShowMore] = useState(false);
@@ -91,29 +56,6 @@ export const FaqSection = () => {
     </Container>
   );
 };
-
-const QuestionContainer = styled(Box)(() => ({
-  paddingTop: "1.063rem",
-  paddingBottom: "1.063rem",
-  borderBottom: "1px solid #D9D9D9",
-}));
-
-const QuestionTitleContainer = styled(Box)(() => ({
-  display: "flex",
-  justifyContent: "space-between",
-  cursor: "pointer",
-}));
-const QuestionTitle = styled(Box)(({ theme }) => ({
-  fontSize: "1rem",
-  fontWeight: 300,
-  color: theme.palette.textPrimary.main,
-}));
-const QuestionAnswer = styled(Box)(({ theme }) => ({
-  paddingTop: "1rem",
-  fontSize: "1rem",
-  fontWeight: 300,
-  color: theme.palette.faqSection.answerColor,
-}));
 
 const FaqQuestion = ({ qna, id }) => {
   const [open, setIsOpen] = useState(false);
