+++ src/components/footer/footer.jsx
@@ -1,5 +1,3 @@
-import { Box, Link, styled } from "@mui/material";
-import Background from "../../assets/images/footer-image.jpg";
 import LogoImage from "../../assets/images/footer-logo.svg";
 
 import facebook from "../../assets/icons/social/facebook.svg";
@@ -7,57 +5,7 @@ import instagram from "../../assets/icons/social/instagram.svg";
 import twitter from "../../assets/icons/social/twitter.svg";
 import youtube from "../../assets/icons/social/youtube.svg";
 
-const Container = styled(Box)(() => ({
-  fontFamily: "'Aquawax', Courier, monospace;",
-  height: 300,
-  boxShadow: "inset 0 0 0 1000px rgba(0,0,0,.5);",
-  backgroundImage: "url(" + Background + ")",
-  backgroundPosition: "center",
-  backgroundRepeat: "no-repeat",
-  backgroundSize: "cover",
-  display: "flex",
-  flexDirection: "column",
-  justifyContent: "center",
-  alignItems: "center",
-  gap: "2.5rem",
-}));
-
-const InnerFooter = styled(Box)(({ theme }) => ({
-  minHeight: 30,
-  fontFamily: "Rubik",
-  background: theme.palette.primary.main,
-  width: "auto",
-  textAlign: "center",
-  color: theme.palette.buttonPrimary.contrastText,
-  fontSize: "1rem",
-  display: "flex",
-  justifyContent: "center",
-  alignItems: "center",
-  padding: "0 1rem 0 1rem",
-  gap: "0.5rem",
-  flexWrap: "wrap",
-}));
-
-const Logo = styled(Box)(() => ({}));
-
-const SocialContainer = styled(Box)(() => ({
-  display: "flex",
-  gap: "3.313rem",
-  padding: "0 1rem 0 1rem",
-  flexWrap: "wrap",
-  justifyContent: "center",
-}));
-
-const SocialLink = styled(Link)(() => ({
-  height: 24,
-  display: "flex",
-  flexDirection: "column",
-  justifyContent: "center",
-  alignItems: "center",
-  color: "white",
-  textDecoration: "none",
-  cursor: "pointer",
-}));
+import { Container, InnerFooter, Logo, SocialContainer, SocialLink } from "./footer.styles";
 
 export const Footer = () => {
   return (
