+++ src/components/howItWorks/howItWorks.jsx
@@ -1,59 +1,16 @@
-import { DialogContent, Dialog, Button, styled, TextField, Box, DialogTitle } from "@mui/material";
+import { Button } from "@mui/material";
 import { useState } from "react";
-import { PrimarySubText } from "../common/typographies/typographies";
-import { PrimaryText } from "../common/typographies/typographies";
 import arrowRight from "../../assets/icons/arrow-right_filled.svg";
 import CloseIcon from "../../assets/icons/Close.svg";
 import { Icon } from "../common/icon/icon";
-const BootstrapDialog = styled(Dialog)(({ theme }) => ({
-  height: "auto",
-  width: "auto",
-}));
-const StyledDialogTitle = styled(DialogTitle)(({ theme }) => ({
-  backgroundColor: theme.palette.primary.main,
-  display: "flex",
-  flexDirection: "row",
-  justifyContent: "center",
-  alignItems: "center",
-  height: 40,
-  padding: 0,
-}));
-
-const CustomeTitle = styled(PrimarySubText)(({ theme }) => ({
-  fontFamily: "rubik",
-  fontWeight: 700,
-  fontSize: 15,
-  textAlign: "center",
-  color: theme.palette.primary.contrastText,
-  position: "relative",
-}));
-const StyledDialogContent = styled(DialogContent)(({ theme }) => ({
-  padding: 0,
-  marginTop: 10,
-  marginLeft: 16,
-  marginRight: 16,
-  marginBottom: 33,
-  "&.MuiDialogContent-root": {
-    overflowY: "unset",
-  },
-}));
-
-const SubText = styled(PrimarySubText)(({ theme }) => ({
-  fontSize: 12,
-  fontFamily: theme.fontFamily.Inter,
-  [theme.breakpoints.down("sm")]: {
-    paddingBottom: 13,
-  },
-}));
-
-const Title = styled(PrimaryText)(({ theme }) => ({
-  fontSize: 15,
-  fontWeight: 700,
-  fontFamily: theme.fontFamily.Rubik,
-  textAlign: "start",
-  paddingTop: 15,
-  paddingBottom: 13,
-}));
+import {
+  BootstrapDialog,
+  StyledDialogTitle,
+  CustomeTitle,
+  StyledDialogContent,
+  SubText,
+  Title,
+} from "./howItWorks.styles";
 
 export const HowItWorks = () => {
   const [open, setOpen] = useState(false);
