+++ src/components/packageDetails/packageDetails.styles.js
@@ -1,0 +1,8 @@
+import { styled, Box } from "@mui/material";
+
+export const PackageDetailsMainBox = styled(Box)(({ theme }) => ({
+  padding: "0px 0px 0px 0px",
+  backgroundColor: "white",
+  border: `2px solid ${theme.palette.primary.main}`,
+  // height: 870
+}));
