+++ src/components/widgets/formInput/formInput.jsx
@@ -1,21 +1,4 @@
-import { Box, TextField, styled } from "@mui/material";
-
-const FormInputBox = styled(Box)(({ theme }) => ({
-  display: "flex",
-  flexDirection: "column",
-  gap: "0.8rem",
-  flex: 1,
-  "& span": {
-    fontFamily: theme.typography.primaryTitle.fontFamily,
-  },
-}));
-
-const FormTextInput = styled(TextField)(() => ({
-  background: "white",
-  border: "none",
-  "& MuiFormControl-root:": {},
-}));
-
+import { FormInputBox, FormTextInput } from "./formInput.styles";
 export const FormInput = ({ label, placeholder, multiline = false, value, onChange, minRows = 4, name }) => {
   return (
     <FormInputBox>
